JBPM-4729 - Add process instance filters based on deploydesc on datasets - exposed deployment by user via dedicated class
