+++ jbpm-services/jbpm-services-cdi/src/main/java/org/jbpm/services/cdi/impl/ExcludeNonCDIImplExtension.java
@@ -42,6 +42,7 @@ public class ExcludeNonCDIImplExtension implements Extension {
 		excluded.add("org.jbpm.services.task.wih.ExternalTaskEventListener");
 		excluded.add("org.jbpm.services.task.events.DefaultTaskEventListener");
 		excluded.add("org.jbpm.kie.services.impl.store.DeploymentSynchronizer");
+		excluded.add("org.jbpm.kie.services.impl.security.DeploymentRolesManager");
 	}
