+++ jbpm-services/jbpm-services-cdi/src/test/java/org/jbpm/services/cdi/impl/manager/SingleRuntimeManagerWithListenersTest.java
@@ -121,6 +121,7 @@ public class SingleRuntimeManagerWithListenersTest extends AbstractKieServicesBa
                 .addPackage("org.jbpm.services.cdi.impl.manager")
                 .addPackage("org.jbpm.services.cdi.producer")
                 .addPackage("org.jbpm.services.cdi.impl.producers")
+                .addPackage("org.jbpm.services.cdi.impl.security")
                 
                 .addPackage("org.jbpm.kie.services.test")
                 .addPackage("org.jbpm.services.cdi.test") // Identity Provider Test Impl here
