+++ jbpm-services/jbpm-services-cdi/src/test/java/org/jbpm/services/cdi/test/DeploymentServiceCDIImplWithSyncTest.java
@@ -94,6 +94,7 @@ public class DeploymentServiceCDIImplWithSyncTest extends DeploymentServiceWithS
                 .addPackage("org.jbpm.services.cdi.impl.manager")
                 .addPackage("org.jbpm.services.cdi.impl.store")
                 .addPackage("org.jbpm.services.cdi.producer")
+                .addPackage("org.jbpm.services.cdi.impl.security")
                 
                 .addPackage("org.jbpm.test.util")
                 .addPackage("org.jbpm.kie.services.test")
