+++ jbpm-services/jbpm-services-cdi/src/test/java/org/jbpm/services/cdi/test/UserTaskServiceCDIImplTest.java
@@ -93,6 +93,7 @@ public class UserTaskServiceCDIImplTest extends UserTaskServiceImplTest {
                 .addPackage("org.jbpm.services.cdi.impl.form")
                 .addPackage("org.jbpm.services.cdi.impl.manager")
                 .addPackage("org.jbpm.services.cdi.producer")
+                .addPackage("org.jbpm.services.cdi.impl.security")
                 
                 .addPackage("org.jbpm.test.util")
                 .addPackage("org.jbpm.kie.services.test")
