+++ jbpm-services/jbpm-services-cdi/src/test/java/org/jbpm/services/cdi/test/support/CDISupportProcessTSInjectionDisabledTest.java
@@ -81,6 +81,7 @@ public class CDISupportProcessTSInjectionDisabledTest extends SupportProcessBase
                  .addPackage("org.jbpm.services.cdi.impl.form")
                  .addPackage("org.jbpm.services.cdi.impl.manager")
                  .addPackage("org.jbpm.services.cdi.producer")
+                 .addPackage("org.jbpm.services.cdi.impl.security")
                  
                  .addPackage("org.jbpm.kie.services.test")
                  .addPackage("org.jbpm.services.cdi.test") // Identity Provider Test Impl here
