+++ jbpm-services/jbpm-services-ejb/jbpm-services-ejb-impl/src/main/java/org/jbpm/services/ejb/impl/security/DeploymentRolesManagerEJBImpl.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2015 JBoss by Red Hat.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jbpm.services.ejb.impl.security;
+
+import java.util.List;
+
+import javax.ejb.ConcurrencyManagement;
+import javax.ejb.ConcurrencyManagementType;
+import javax.ejb.Lock;
+import javax.ejb.LockType;
+import javax.ejb.Singleton;
+
+import org.jbpm.kie.services.impl.security.DeploymentRolesManager;
+import org.kie.internal.identity.IdentityProvider;
+
+@Singleton
+@ConcurrencyManagement(ConcurrencyManagementType.CONTAINER)
+@Lock(LockType.READ)
+public class DeploymentRolesManagerEJBImpl extends DeploymentRolesManager {
+
+    @Lock(LockType.WRITE)
+    @Override
+    public void addRolesForDeployment(String deploymentId, List<String> roles) {
+        super.addRolesForDeployment(deploymentId, roles);
+    }
+
+    @Lock(LockType.WRITE)
+    @Override
+    public void removeRolesForDeployment(String deploymentId) {
+        super.removeRolesForDeployment(deploymentId);
+    }
+
+    @Lock(LockType.READ)
+    @Override
+    public List<String> getDeploymentsForUser(IdentityProvider identityProvider) {
+        return super.getDeploymentsForUser(identityProvider);
+    }
+
+}
